Completing the IntCode Computer to allow for relative instruction mode

+++ Advent2019/Solutions/Day07.cs
@@ -6,7 +6,7 @@ namespace Advent2019.Solutions
 {
     class Day07
     {
-        public int RunAmplifiers(int[] phases, string program)
+        public long RunAmplifiers(int[] phases, string program)
         {
             var amplifiers = BuildAmplifiers(phases, program);
             amplifiers.ForEach(x => x.RunProgram());
@@ -14,7 +14,7 @@ namespace Advent2019.Solutions
             return amplifiers.Last().Outputs.Single();
         }
 
-        public int FindLargestSignal(string program)
+        public long FindLargestSignal(string program)
         {
             return GetPermutations(new[] {0, 1, 2, 3, 4}, 5)
                 .Select(x=> RunAmplifiers(x.ToArray(), program))
@@ -51,7 +51,7 @@ namespace Advent2019.Solutions
             return amplifiers;
         }
 
-        public async Task<int> RunFeedbackLoopAsync(int[] phases, string program)
+        public async Task<long> RunFeedbackLoopAsync(int[] phases, string program)
         {
             var amplifiers = BuildAmplifiers(phases, program, true);
 
@@ -62,7 +62,7 @@ namespace Advent2019.Solutions
             return amplifiers.Last().Outputs.First();
         }
 
-        public int FindLargestSignalWithFeedbackLoop(string program)
+        public long FindLargestSignalWithFeedbackLoop(string program)
         {
             var tasks = GetPermutations(new[] {9, 8, 7, 6, 5}, 5)
                 .Select(async x => await RunFeedbackLoopAsync(x.ToArray(), program));
